+++ src/surrealist/ql/database.py
@@ -1,15 +1,13 @@
 import logging
-import warnings
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 from surrealist.connections.connection import Connection
-from surrealist.enums import Algorithm, AutoOrNone
+from surrealist.enums import AutoOrNone
 from surrealist.errors import SurrealConnectionError
 from surrealist.ql.statements import Access, Live, Remove, Select
 from surrealist.ql.statements.alter import Alter
 from surrealist.ql.statements.define import (DefineEvent, DefineIndex,
-                                             DefineParam, DefineScope,
-                                             DefineTable, DefineToken)
+                                             DefineParam, DefineTable)
 from surrealist.ql.statements.define_access import (DefineAccessBearer,
                                                     DefineAccessJwt,
                                                     DefineAccessRecord)
@@ -310,44 +308,6 @@ class Database:
         """
         return Remove(self._connection, '', type_="ANALYZER", name=name)
 
-    def define_scope(self, name: str, duration: str, signup: Union[str, Statement],
-                     signin: Union[str, Statement]) -> DefineScope:
-        """
-        Deprecated since SurrealDB 2.x, use define_access_record instead!
-
-        Represents DEFINE SCOPE statement
-
-        Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/scope
-
-        Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
-
-        :param name: name for the new scope
-        :param duration: session duration, like 24h
-        :param signup: Create statement with string or Statement representation
-        :param signin: Select statement with string or Statement representation
-        :return: DefineScope object
-        """
-        msg = "Deprecated since SurrealDB 2.x, use define_access_record instead\n" \
-              "Read more here: https://surrealdb.com/docs/surrealql/statements/define/access"
-        warnings.warn(msg)
-        logger.warning(msg)
-        return DefineScope(self._connection, name, duration, signup, signin)
-
-    def remove_scope(self, name: str) -> Remove:
-        """
-        Do not work since SurrealDB 2.x, use remove_access instead!
-
-        Remove the scope
-
-        :param name: name of the scope
-        :return: Remove object
-        """
-        msg = "Do not work since SurrealDB 2.x, use remove_access instead\n" \
-              "Read more here: https://surrealdb.com/docs/surrealql/statements/remove"
-        warnings.warn(msg)
-        logger.warning(msg)
-        return Remove(self._connection, "", type_="SCOPE", name=name)
-
     def define_index(self, name: str, table_name: str) -> DefineIndex:
         """
         Represents DEFINE INDEX statement
@@ -366,7 +326,7 @@ class Database:
         """
         Represents REBUILD INDEX object, used to rebuild resources.
 
-        Refer to: https://surrealdb.com/docs/surrealdb/surrealql/statements/rebuild
+        Refer to: https://surrealdb.com/docs/surrealql/statements/rebuild
 
         :param index_name: name of the index
         :param table_name: name of the table
@@ -385,42 +345,6 @@ class Database:
         """
         return Remove(self._connection, name=name, table_name=table_name, type_="INDEX")
 
-    def define_token(self, name: str, token_type: Algorithm, value: str) -> DefineToken:
-        """
-        Deprecated since SurrealDB 2.x, use define_access_jwt or define_access_record instead!
-
-        Represents DEFINE TOKEN statement
-
-        Refer to: https://docs.surrealdb.com/docs/surrealql/statements/define/token
-
-        Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/database.py
-
-        :param name: name for the token
-        :param token_type: type of the token, for example, Algorithm.RS256
-        :param value: value of the token
-        :return: DefineToken object
-        """
-        msg = "Deprecated since SurrealDB 2.x, use define_access_jwt or define_access_record instead\n" \
-              "Read more here: https://surrealdb.com/docs/surrealql/statements/define/access"
-        warnings.warn(msg)
-        logger.warning(msg)
-        return DefineToken(self._connection, name, token_type, value)
-
-    def remove_token(self, name: str) -> Remove:
-        """
-        Do not work since SurrealDB 2.x, use remove_access instead!
-
-        Remove token by name for the database
-
-        :param name: name of the token
-        :return: Remove object
-        """
-        msg = "Do not work since SurrealDB 2.x, use remove_access instead\n" \
-              "Read more here: https://surrealdb.com/docs/surrealql/statements/remove"
-        warnings.warn(msg)
-        logger.warning(msg)
-        return Remove(self._connection, "", type_="TOKEN", name=name)
-
     def define_access_jwt(self, name: str) -> DefineAccessJwt:
         """
         Represents DEFINE ACCESS ... JWT statement.
@@ -438,7 +362,6 @@ class Database:
     def define_access_record(self, name: str) -> DefineAccessRecord:
         """
         Represents DEFINE ACCESS ... RECORD statement.
-        Use this method instead of define_token or define_scope
 
         Refer to: https://surrealdb.com/docs/surrealql/statements/define/access/record
 
@@ -493,7 +416,7 @@ class Database:
         Example:
         db.live_query("person", func).alias("first_name", "NAME").where("age > 22").run()
 
-        Refer to: https://surrealdb.com/docs/surrealdb/surrealql/statements/live
+        Refer to: https://surrealdb.com/docs/surrealql/statements/live
 
         Refer to: https://github.com/kotolex/surrealist?tab=readme-ov-file#live-query
 
@@ -596,7 +519,7 @@ class Database:
         """
         Represents ALTER TABLE statement
 
-        Refer to: https://surrealdb.com/docs/surrealdb/surrealql/statements/alter
+        Refer to: https://surrealdb.com/docs/surrealql/statements/alter
 
         Example: https://github.com/kotolex/surrealist/blob/master/examples/surreal_ql/ql_alter_examples.py